+++ packages/openapi3/test/discriminator.test.ts
@@ -1,4 +1,4 @@
-import { ModelType, ModelTypeProperty } from "@cadl-lang/compiler";
+import { Model, ModelProperty } from "@cadl-lang/compiler";
 import { expectDiagnostics } from "@cadl-lang/compiler/testing";
 import { deepStrictEqual, match, ok, strictEqual } from "assert";
 import { checkFor, createOpenAPITestRunner, openApiFor } from "./test-host.js";
@@ -353,13 +353,13 @@ describe("openapi3: discriminated unions", () => {
       op read(): { @body body: Pet };
       `);
     strictEqual(diagnostics.length, 6);
-    strictEqual((diagnostics[0].target as ModelType).name, "Dog");
+    strictEqual((diagnostics[0].target as Model).name, "Dog");
     match(diagnostics[0].message, /not defined in a variant of a discriminated union/);
-    strictEqual((diagnostics[1].target as ModelTypeProperty).name, "kind"); // Pig.kind
+    strictEqual((diagnostics[1].target as ModelProperty).name, "kind"); // Pig.kind
     match(diagnostics[1].message, /must be type 'string'/);
-    strictEqual((diagnostics[2].target as ModelTypeProperty).name, "kind"); // Tiger.kind
+    strictEqual((diagnostics[2].target as ModelProperty).name, "kind"); // Tiger.kind
     match(diagnostics[2].message, /must be a required property/);
-    strictEqual((diagnostics[3].target as ModelType).name, "Dog");
+    strictEqual((diagnostics[3].target as Model).name, "Dog");
     match(diagnostics[3].message, /define the discriminator property with a string literal value/);
     match(diagnostics[4].message, /define the discriminator property with a string literal value/); // Pig.kind
     match(diagnostics[5].message, /define the discriminator property with a string literal value/); // Lizard.kind