+++ packages/openapi3/test/versioning.test.ts
@@ -1,5 +1,7 @@
+import { DecoratorContext, Namespace } from "@cadl-lang/compiler";
+import { createTestWrapper } from "@cadl-lang/compiler/testing";
 import { deepStrictEqual, strictEqual } from "assert";
-import { openApiFor } from "./test-host.js";
+import { createOpenAPITestHost, createOpenAPITestRunner, openApiFor } from "./test-host.js";
 
 describe("openapi3: versioning", () => {
   it("works with models", async () => {
@@ -98,5 +100,79 @@ describe("openapi3: versioning", () => {
       },
       required: ["prop1", "prop2", "prop3"],
     });
+  });
+
+  it("doesn't lose parent namespace", async () => {
+    const host = await createOpenAPITestHost();
+
+    let storedNamespace: string | undefined = undefined;
+    host.addJsFile("test.js", {
+      $armNamespace(context: DecoratorContext, entity: Namespace) {
+        storedNamespace = context.program.checker.getNamespaceString(entity);
+      },
+    });
+
+    const runner = createTestWrapper(
+      host,
+      (code) =>
+        `import "@cadl-lang/rest"; import "@cadl-lang/openapi";
+          import "@cadl-lang/openapi3"; import "@cadl-lang/versioning";
+          import "./test.js";
+          using Cadl.Rest; using Cadl.Http; using OpenAPI; using Cadl.Versioning; ${code}`,
+      { emitters: { "@cadl-lang/openapi3": {} } }
+    );
+
+    await runner.compile(`
+    @versioned(Contoso.Library.Versions)
+    namespace Contoso.Library {
+      namespace Blah { }
+      enum Versions { v1 };
+    }
+    @armNamespace
+    @serviceTitle("Widgets 'r' Us")
+    @versionedDependency(Contoso.Library.Versions.v1)
+    namespace Contoso.WidgetService {
+      model Widget {
+        @key
+        @segment("widgets")
+        id: string;
+      }
+      interface Operations {
+        @test
+        op get(id: string): Widget;
+      }
+    }
+    `);
+
+    strictEqual(storedNamespace, "Contoso.WidgetService");
+  });
+
+  // Test for https://github.com/microsoft/cadl/issues/812
+  it("doesn't throw errors when using UpdateableProperties", async () => {
+    // if this test throws a duplicate name diagnostic, check that getEffectiveType
+    // is returning the projected type.
+    const runner = await createOpenAPITestRunner({ withVersioning: true });
+    await runner.compile(`
+      @versioned(Library.Versions)
+      namespace Library {
+        enum Versions {
+          v1,
+          v2,
+        }
+      }
+      
+      @serviceTitle("Service")
+      @versionedDependency(Library.Versions.v1)
+      namespace Service {
+        model Widget {
+          details?: WidgetDetails;
+        }
+      
+        model WidgetDetails {}
+        interface Projects {
+          oops(...UpdateableProperties<Widget>): Widget;
+        }
+      }
+    `);
   });
 });